Move committed block processor
