+++ packages/consensus/source/processors/abstract-processor.ts
@@ -4,17 +4,13 @@ import { Contracts, Identifiers } from "@mainsail/contracts";
 @injectable()
 export class AbstractProcessor {
 	@inject(Identifiers.Application)
-	private readonly app!: Contracts.Kernel.Application;
+	protected readonly app!: Contracts.Kernel.Application;
 
 	protected hasValidHeightOrRound(message: { height: number; round: number }): boolean {
-		return message.height === this.#getConsensus().getHeight() && message.round >= this.#getConsensus().getRound();
+		return message.height === this.getConsensus().getHeight() && message.round >= this.getConsensus().getRound();
 	}
 
-	protected handle(roundState: Contracts.Consensus.IRoundState) {
-		void this.#getConsensus().handle(roundState);
-	}
-
-	#getConsensus(): Contracts.Consensus.IConsensusService {
+	protected getConsensus(): Contracts.Consensus.IConsensusService {
 		return this.app.get<Contracts.Consensus.IConsensusService>(Identifiers.Consensus.Service);
 	}
 }
